pyproject.toml
[tool.poetry]
name = "term-thesis"
version = "0.1.0"
description = ""
authors = ["Al_Pl <asuspitsyn@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
openpyxl = "^3.1.4"
pyxlsb = "^1.0.10"
pytest = "^8.2.2"
python-dotenv = "^1.0.1"
any = "^0.0.2"
optional = "^0.0.1"
xlrd = "^2.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '.git'

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = "venv/*"
